Move ILCDBase and ILCD into another file
